1.20.2 (#610)
* First stab at 1.20.2

* Improve smooth teleport

Revert much of the reflection

* Fix build to include new version

* Update IWorldHandler.java
