+++ .github/workflows/maven.yml
@@ -125,52 +125,52 @@ jobs:
         if: steps.wild.outputs.sucess != 'true' || steps.wildMojang.outputs.sucess != 'true' || steps.wildObf.outputs.sucess != 'true'
         run: cd BuildTools && java -jar BuildTools.jar --rev 1.19.4 --remapped
 
-  # Build 1.20.1 NMS
-  v1_20_R1:
+  # Build 1.20.2 NMS
+  v1_20_R2:
     runs-on: ubuntu-latest
     steps:
-      - name: Set up JDK 17 # 1.20.1 can only be built with Java 17
+      - name: Set up JDK 17 # 1.20.2 can only be built with Java 17
         uses: actions/setup-java@v3
         with:
           distribution: 'temurin'
           java-version: '17'
-      - name: Cache 1.20.1 Maven package
+      - name: Cache 1.20.2 Maven package
         id: cacheWild_r2
         uses: actions/cache@v3
         with:
           path: |
-            ~/.m2/repository/org/spigotmc/spigot/1.20.1-R0.1-SNAPSHOT/
+            ~/.m2/repository/org/spigotmc/spigot/1.20.2-R0.1-SNAPSHOT/
             ~/.m2/repository/org/spigotmc/spigot-parent/
             ~/.m2/repository/org/spigotmc/minecraft-server/
-          key: ${{ runner.os }}-v1_20_R1
-          restore-keys: ${{ runner.os }}-v1_20_R1
+          key: ${{ runner.os }}-v1_20_R2
+          restore-keys: ${{ runner.os }}-v1_20_R2
       - name: Cache Maven packages
         id: cacheMain
         uses: actions/cache@v3
         with:
           path: ~/.m2
-          key: ${{ runner.os }}-m2-v1_20_R1
-          restore-keys: ${{ runner.os }}-m2-v1_20_R1
+          key: ${{ runner.os }}-m2-v1_20_R2
+          restore-keys: ${{ runner.os }}-m2-v1_20_R2
 
       - name: Setup BuildTools
         run: mkdir BuildTools && wget -O BuildTools/BuildTools.jar https://hub.spigotmc.org/jenkins/job/BuildTools/lastSuccessfulBuild/artifact/target/BuildTools.jar
-      - name: Check 1.20.1 Spigot
+      - name: Check 1.20.2 Spigot
         id: wild
-        run: test -f ~/.m2/repository/org/spigotmc/spigot/1.20.1-R0.1-SNAPSHOT/spigot-1.20.1-R0.1-SNAPSHOT.jar && echo "sucess=true" >> $GITHUB_OUTPUT || echo "sucess=false" >> $GITHUB_OUTPUT
-      - name: Check 1.20.1 Spigot (Mojang)
+        run: test -f ~/.m2/repository/org/spigotmc/spigot/1.20.2-R0.1-SNAPSHOT/spigot-1.20.2-R0.1-SNAPSHOT.jar && echo "sucess=true" >> $GITHUB_OUTPUT || echo "sucess=false" >> $GITHUB_OUTPUT
+      - name: Check 1.20.2 Spigot (Mojang)
         id: wildMojang
-        run: test -f ~/.m2/repository/org/spigotmc/spigot/1.20.1-R0.1-SNAPSHOT/spigot-1.20.1-R0.1-SNAPSHOT-remapped-mojang.jar && echo "sucess=true" >> $GITHUB_OUTPUT || echo "sucess=false" >> $GITHUB_OUTPUT
-      - name: Check 1.20.1 Spigot (Obf)
+        run: test -f ~/.m2/repository/org/spigotmc/spigot/1.20.2-R0.1-SNAPSHOT/spigot-1.20.2-R0.1-SNAPSHOT-remapped-mojang.jar && echo "sucess=true" >> $GITHUB_OUTPUT || echo "sucess=false" >> $GITHUB_OUTPUT
+      - name: Check 1.20.2 Spigot (Obf)
         id: wildObf
-        run: test -f ~/.m2/repository/org/spigotmc/spigot/1.20.1-R0.1-SNAPSHOT/spigot-1.20.1-R0.1-SNAPSHOT-remapped-obf.jar && echo "sucess=true" >> $GITHUB_OUTPUT || echo "sucess=false" >> $GITHUB_OUTPUT
-      - name: Build 1.20.1
+        run: test -f ~/.m2/repository/org/spigotmc/spigot/1.20.2-R0.1-SNAPSHOT/spigot-1.20.2-R0.1-SNAPSHOT-remapped-obf.jar && echo "sucess=true" >> $GITHUB_OUTPUT || echo "sucess=false" >> $GITHUB_OUTPUT
+      - name: Build 1.20.2
         if: steps.wild.outputs.sucess != 'true' || steps.wildMojang.outputs.sucess != 'true' || steps.wildObf.outputs.sucess != 'true'
-        run: cd BuildTools && java -jar BuildTools.jar --rev 1.20.1 --remapped
+        run: cd BuildTools && java -jar BuildTools.jar --rev 1.20.2 --remapped
 
   # Build Movecraft
   build:
     runs-on: ubuntu-latest
-    needs: [v1_16_R3, v1_18_R2, v1_19_R3, v1_20_R1]
+    needs: [v1_16_R3, v1_18_R2, v1_19_R3, v1_20_R2]
 
     steps:
       - name: Checkout Movecraft
@@ -210,15 +210,15 @@ jobs:
             ~/.m2/repository/org/spigotmc/minecraft-server/
           key: ${{ runner.os }}-v1_19_R3
           restore-keys: ${{ runner.os }}-v1_19_R3
-      - name: Cache 1.20.1 Maven package
+      - name: Cache 1.20.2 Maven package
         uses: actions/cache@v3
         with:
           path: |
-            ~/.m2/repository/org/spigotmc/spigot/1.20.1-R0.1-SNAPSHOT/
+            ~/.m2/repository/org/spigotmc/spigot/1.20.2-R0.1-SNAPSHOT/
             ~/.m2/repository/org/spigotmc/spigot-parent/
             ~/.m2/repository/org/spigotmc/minecraft-server/
-          key: ${{ runner.os }}-v1_20_R1
-          restore-keys: ${{ runner.os }}-v1_20_R1
+          key: ${{ runner.os }}-v1_20_R2
+          restore-keys: ${{ runner.os }}-v1_20_R2
 
       - name: Build with Maven
         run: mvn -T 1C -B package --file pom.xml
