+++ modules/Movecraft/pom.xml
@@ -50,7 +50,7 @@
         </dependency>
         <dependency>
             <groupId>net.countercraft</groupId>
-            <artifactId>movecraft-v1_20_r1</artifactId>
+            <artifactId>movecraft-v1_20_r2</artifactId>
             <version>${revision}</version>
             <type>jar</type>
         </dependency>
@@ -190,7 +190,7 @@
                                     </includes>
                                 </filter>
                                 <filter>
-                                    <artifact>net.countercraft:movecraft-v1_20_r1</artifact>
+                                    <artifact>net.countercraft:movecraft-v1_20_r2</artifact>
                                     <includes>
                                         <include>**</include>
                                     </includes>
