+++ modules/v1_20_R2/src/main/java/net/countercraft/movecraft/compat/v1_20_R2/NextTickProvider.java
@@ -1,4 +1,4 @@
-package net.countercraft.movecraft.compat.v1_20_R1;
+package net.countercraft.movecraft.compat.v1_20_R2;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
