+++ pom.xml
@@ -17,7 +17,7 @@
         <module>modules/v1_16_R3</module>
         <module>modules/v1_18_R2</module>
         <module>modules/v1_19_R3</module>
-        <module>modules/v1_20_R1</module>
+        <module>modules/v1_20_R2</module>
         <module>modules/datapack</module>
         <module>modules/Movecraft</module>
     </modules>
