+++ setup.sh
@@ -47,11 +47,11 @@ else
     java -jar BuildTools.jar --rev 1.19.4 --remapped
 fi
 
-# Build 1.20.1
-if [ -f ~/.m2/repository/org/spigotmc/spigot/1.20.1-R0.1-SNAPSHOT/spigot-1.20.1-R0.1-SNAPSHOT.jar ]; then
-    echo "1.20.1 already exists, skipping build"
+# Build 1.20.2
+if [ -f ~/.m2/repository/org/spigotmc/spigot/1.20.2-R0.1-SNAPSHOT/spigot-1.20.2-R0.1-SNAPSHOT.jar ]; then
+    echo "1.20.2 already exists, skipping build"
 else
-    java -jar BuildTools.jar --rev 1.20.1 --remapped
+    java -jar BuildTools.jar --rev 1.20.2 --remapped
 fi
 
 # Restore git information
